Got Base loop working but combat not fully functional

+++ Player.cpp
@@ -6,7 +6,7 @@ namespace Play
 {
 	int Player::playerArmor;
 	int Player::playerWeapon;
-	float Player::playerHp;
-	float Player::playerAtk;
+	int Player::playerHp;
+	int Player::playerAtk;
 	vector<unique_ptr<Item>> playerItems;
 }
